interest_analysis/requirements.txt
python-dotenv

pydantic

requests 

fastapi
uvicorn

openai
langchain 
langchain-openai
langchain-community

pymongo